Move cacheCurrentUser to user-utils
